+++ Dfe.Academies.Academisation.Service.UnitTest/Commands/TransferProject/SetTransferProjectTransferDatesCommandHandlerTests.cs
@@ -62,7 +62,8 @@ namespace Dfe.Academies.Academisation.Service.UnitTest.Commands.TransferProject
 				TargetDateForTransfer = DateTime.UtcNow.AddMonths(1)
 			};
 			// Create a transfer project to 'SetTransferDates' to
-			var transferProject = Domain.TransferProjectAggregate.TransferProject.Create("12345678", "out trust", "23456789", "in trust", new List<string> { "34567890" }, false, DateTime.Now);
+			var transferringAcademies = new List<TransferringAcademy>() { new TransferringAcademy("23456789", "in trust", "34567890", "", "") };
+			var transferProject = Domain.TransferProjectAggregate.TransferProject.Create("12345678", "out trust", transferringAcademies, false, DateTime.Now);
 			// Mock Unit of work and Repository 
 			var unitOfWorkMock = new Mock<IUnitOfWork>();
 			unitOfWorkMock.Setup(u => u.SaveChangesAsync(It.IsAny<CancellationToken>()))