+++ release-notes.md
@@ -1,5 +1,6 @@
 ## NEXT
 * Added ability to edit PFI schemes within a Transfer project for their requisite Transferring Academies
+* Added Region and Local authority to the transfering academy record for filtering purposes
 ## 9.2.0
 * Added transfer projects
 * Added fields required to a Conversion Project to allow for sponsored grants